Refactor to abstract triggerDownload function for laporan
